requirements.txt
google-cloud-bigquery==3.3.0
google-cloud-pubsub==2.13.4
Jinja2==3.1.2